stack/collector/requirements.txt
paho-mqtt>=2.0.0
psycopg[binary]>=3.1.0
